Extract types outside of llms, small change
